New post on alerting on process completion, and some site changes. (#18)
* New post on alerting on process completion, and some site changes.

The look of the website has been updated with a new color-palette. It's not set in stone but something I'm just trying out. Might/shall tweak it in the future.

LinkedIn and Twitter icons have been added to the menu bar. A commento server has been set up and the blog has a comment integration.

* Formatting.

* Second draft.

* Update 2023-03-28-alerting-on-process-complete.md

Final draft.